+++ internal/inomap/qino.go
@@ -1,0 +1,25 @@
+package inomap
+
+import (
+	"syscall"
+)
+
+// QIno = Qualified Inode number.
+// Uniquely identifies a backing file through the device number,
+// inode number pair.
+type QIno struct {
+	// Stat_t.{Dev,Ino} is uint64 on 32- and 64-bit Linux
+	Dev uint64
+	Ino uint64
+}
+
+// QInoFromStat fills a new QIno struct with the passed Stat_t info.
+func QInoFromStat(st *syscall.Stat_t) QIno {
+	return QIno{
+		// There are some architectures that use 32-bit values here
+		// (darwin, freebsd-32, maybe others). Add and explicit cast to make
+		// this function work everywhere.
+		Dev: uint64(st.Dev),
+		Ino: uint64(st.Ino),
+	}
+}
